setting login logout & user Info
