+++ server/routes/admin.js
@@ -2,11 +2,10 @@ var express = require('express');
 var router = express.Router();
 
 const adminController = require('../controllers/adminController');
-const publicController = require('../controllers/publicController');
 
-router.get('/login', publicController.viewLogin );
-router.post('/login', publicController.actionLogin );
-router.post('/logout', publicController.actionLogout );
+router.get('/login', adminController.viewLogin );
+router.post('/login', adminController.actionLogin );
+router.get('/logout', adminController.actionLogout );
