+++ server/routes/index.js
@@ -1,10 +1,10 @@
 var express = require('express');
-const publicController = require('../controllers/publicController');
+const adminController = require('../controllers/adminController');
 var router = express.Router();
 /* GET home page. */
 router.get('/', function (req, res, next) {
   res.redirect('/admin/login')
 });
-router.get('/404', publicController.view404 );
+router.get('/404', adminController.view404 );
 
 module.exports = router;
